ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paddlepaddle==2.5.2
paddleocr==2.7.3
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-multipart==0.0.6
httpx==0.25.2